Thêm các API lấy thống kê tư vấn viên theo khung thời gian: câu hỏi đã xóa, câu trả lời đã đưa ra, câu trả lời cần duyệt, lịch tư vấn, bài đăng đã duyệt, và cuộc hội thoại

+++ src/main/java/studentConsulting/model/payload/dto/UserStatisticsDTO.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserStatisticsDTO {
-	private long totalQuestions;
-	private long questionsOver500Views;
-	private long totalAppointments;
-	private long totalRatings;
+	private Integer totalQuestions;
+	private Integer questionsOver500Views;
+	private Integer totalAppointments;
+	private Integer totalRatings;
 }
